+++ besu/src/main/java/org/hyperledger/besu/cli/BesuCommand.java
@@ -1268,10 +1268,18 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
       validateOptions();
       configure();
       initController();
+
+      besuPluginContext.beforeExternalServices();
+
+      var runner = buildRunner();
+      runner.startExternalServices();
+
       startPlugins();
       validatePluginOptions();
       preSynchronization();
-      startSynchronization();
+
+      runner.startEthereumMainLoop();
+      runner.awaitStop();
 
     } catch (final Exception e) {
       throw new ParameterException(this.commandLine, e.getMessage(), e);
@@ -1432,8 +1440,8 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
     preSynchronizationTaskRunner.runTasks(besuController);
   }
 
-  private void startSynchronization() {
-    synchronize(
+  private Runner buildRunner() {
+    return synchronize(
         besuController,
         p2pEnabled,
         p2pTLSConfiguration,
@@ -2453,7 +2461,7 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
   }
 
   // Blockchain synchronization from peers.
-  private void synchronize(
+  private Runner synchronize(
       final BesuController controller,
       final boolean p2pEnabled,
       final Optional<TLSConfiguration> p2pTLSConfiguration,
@@ -2519,8 +2527,8 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
             .build();
 
     addShutdownHook(runner);
-    runner.start();
-    runner.awaitStop();
+
+    return runner;
   }
 
   protected Vertx createVertx(final VertxOptions vertxOptions) {