+++ CHANGELOG.md
@@ -26,6 +26,7 @@ https://hyperledger.jfrog.io/artifactory/besu-binaries/besu/21.10.5/besu-21.10.5
 ### Additions and Improvements
 - Represent baseFee as Wei instead of long accordingly to the spec [#2785](https://github.com/hyperledger/besu/issues/2785)
 - Implements [EIP-4399](https://eips.ethereum.org/EIPS/eip-4399) to repurpose DIFFICULTY opcode after the merge as a source of entropy from the Beacon chain. [#3081](https://github.com/hyperledger/besu/issues/3081)
+- Re-order external services (e.g JsonRpcHttpService) to start before blocks start processing [#3118](https://github.com/hyperledger/besu/pull/3118)
 
 ### Bug Fixes
 - Update log4j to 2.16.0.